// packages/sdk-core/src/generated/protobuf/features_mixer.ts
export enum Mode {
  AUTO = 0,
  MANUAL = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface Config {
  mode: Mode;
  outputs: string[];
  sources: Receiver_Source[];
}
// ...
export const Config = {
  encode(message: Config, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.mode !== 0) {
      writer.uint32(8).int32(message.mode);
    }
    for (const v of message.outputs) {
      writer.uint32(18).string(v!);
    }
    for (const v of message.sources) {
      Receiver_Source.encode(v!, writer.uint32(26).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Config {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseConfig();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.mode = reader.int32() as any;
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.outputs.push(reader.string());
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.sources.push(Receiver_Source.decode(reader, reader.uint32()));
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Config {
    return {
      mode: isSet(object.mode) ? modeFromJSON(object.mode) : 0,
      outputs: globalThis.Array.isArray(object?.outputs) ? object.outputs.map((e: any) => globalThis.String(e)) : [],
      sources: globalThis.Array.isArray(object?.sources)
        ? object.sources.map((e: any) => Receiver_Source.fromJSON(e))
        : [],
    };
  },

  toJSON(message: Config): unknown {
    const obj: any = {};
    if (message.mode !== 0) {
      obj.mode = modeToJSON(message.mode);
    }
    if (message.outputs?.length) {
      obj.outputs = message.outputs;
    }
    if (message.sources?.length) {
      obj.sources = message.sources.map((e) => Receiver_Source.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Config>, I>>(base?: I): Config {
    return Config.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Config>, I>>(object: I): Config {
    const message = createBaseConfig();
    message.mode = object.mode ?? 0;
    message.outputs = object.outputs?.map((e) => e) || [];
    message.sources = object.sources?.map((e) => Receiver_Source.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };